requirements.txt
aiohttp~=3.10.8
fastapi~=0.115.0
httpx~=0.27.2
nltk~=3.9.1
openai~=1.46.1
pydantic~=2.9.2
pypdf~=5.0.0
python-dotenv~=1.0.1
python-multipart~=0.0.12
requests~=2.32.3
starlette~=0.38.5
uvicorn~=0.30.6